+++ packages/adminui-vue/components/controls/USelectEntity.vue
@@ -276,6 +276,15 @@ export default {
     buildAddNewConfig: {
       type: Function,
       default: config => config
+    },
+
+    /**
+     * Search request condition
+     */
+    searchStrategy: {
+      type: String,
+      default: 'like',
+      validator: value => ['like', 'startsWith'].includes(value)
     }
   },
 
@@ -397,7 +406,7 @@ export default {
       this.pageNum = pageNum
 
       const data = await this.repository()
-        .whereIf(query, this.displayAttribute, 'like', query)
+        .whereIf(query, this.displayAttribute, this.searchStrategy, query)
         .start(pageNum * this.pageSize)
         .limit(this.pageSize + 1)
         .select()